[waveviewer] don't scroll ourselves, let the scrollbar do it
There is a visible difference in how GTK+ is scrolling and how we do it.
It's much smoother in GTK+, so lets propate the scroll event from our
window to the scrollbar and the rest is done by GTK+.

Maybe something similar could be done for the button press events.