[waveviewer] fix view after text direction change
In the rare case that the text direction is changed while the program is
running, the view and cursor didn't update properly. Everything was
shown in the new direction but not the same view (part of the waveform).
Make sure we are showing the same portion of the waveform after the change.