[waveviewer] add new property "zoom"
Set this to TRUE to indicate that the next wave data will be the same
waveform but at a different zoom level. PtWaveviewer needs this to
decide which part of the waveform should be shown.

Note: This will be probably removed with the next API break in
version 1.6 and integrated into pt_waveviewer_set_wave().