[window] fix "insert" action sensitivity on Wayland
With commit 20d8f8dd221ff5c89ae17c9f55c87af56c969997 "insert timestamp" is
only active, if there is something to insert.

Some displays don't notify of changes in the clipboard, this can be checked
with gdk_display_supports_selection_notification(). Wayland returns FALSE
for that (doesn't support it) but obviously it does emit "owner-change"
signals. Either this is a bug or I don't fully understand the concept.

The result was an always active "insert" action, even with no waveform
loaded. With this commit we do that check only on X11.