[player] add property "pause" and getter, and a "play and rewind" function
New symbols:
* pt_player_pause_and_rewind
* pt_player_get_pause

First step to move code away from library consumers to the library. The
reasoning is to make it easier to make plugins for other programs as they
would need less code.

The other goal is to make it possible to fully control the player without
having to control the GUI. For example mediakeys would control the player
directly, the GUI should update itself accordingly.