Add --disable-rpath to configure.ac
This adds a `--disable-rpath` option to the configure script for both
the root directory and libparlatype, causing it to patch the generated
libtool so that it will not set RPATH in the resulting binaries.

This is necessary for packaging on some Linux distributions which
prohibit the use of RPATH in packaged binaries/libraries.