[window] remove non-functional drag'n'drop from time label
As long as the time was shown as a GtkLabel, it could be used as a
source for drag'n'drop operations. This doesn't seem to work with the
label being now on a button. This is a regression of the new design
and I don't see a way to circumvent this, hence removing the code
and stop mentioning it in the help pages.