[player] remove failable object initialization interface from PtPlayer
Make it an ordinary GObject that doesn't fail during initialization.
A new PtPlayer is not usable now, it has to be set up first with
pt_player_setup_player(). This setup function can fail with a GError.

The reasoning is to allow different GStreamer pipelines which can be
set up separately. At the moment only a playback pipeline is implemented.

On the application side there is no test of the pipeline before starting
the application anymore. The application will quit if
pt_player_setup_player() returns an error.