[waveloader] load into memory on-the-fly
Load wave data into memory instead of writing it to a temporary file.

Previously raw data was written to disk with the fdsink plugin. This has
the disadvantage that we have to wait until the operation is finished.

The new solution is using the appsink plugin to load the raw data to
memory (into a GArray). To reduce memory usage the audio data is
downsampled to a sample rate of 8000. With 16 bit per sample we need
roughly 1 MB per minute.

This internal, high resolution array is downsampled to the requested
resolution of 50 to 200 min/max values per second. The implementation has
changed to allow all values (previously the requested value could have
been changed for easier computing).

Wave data is now provided on-the-fly: As the audio stream is read in, data
in the GArray is constantly updated. PtWaveloader provides a pointer to the
array, callers shall not modify the data.

This and the following commits serve also as an example of the bad
practice of making huge commits. The original commit history was too
twisted to follow.