[waveviewer] use private waveloader
Add a private PtWaveloader to load and provide wave data. Add an API for
applications to load a wave form.

Add a "progress" signal for applications to show loading progress.

Add a "pps" property (pixels per second) to change the resolution of the
shown wave.

Remove the "zoom" property, which is not necessary anymore.