Remove ASR output handling, rely on D-Bus signals instead
With the previous commit we are emitting D-Bus signals with recognized
strings. Applications can listen to those signals and handle them
themselves, like Parlatype LibreOffice Extension does.

Remove all code that tries to detect possible target applications via
ATSPI and "injects" the strings. Although this was an interesting and
promising approach, it became too complicated to handle, moreover
assistive technologies will probably change with GTK 4.

Remove the "-t" or "--textpad" commandline option, a fallback solution
when ATSPI is not available. It was used by default on Flatpak; remove
code to detect and handle Flatpak as well.

Remove dependency on atspi-2.