[window] remove not working functions
Remove 2 functions which were supposed to give user feedback that a jump
action has been performed (from "outside", e.g. Mediakeys). They clicked
the jump backward/forward button to give visible feedback, however, the
button is either not pressed visibly in or this is done so fast that you
can't actually see it.

The signals PtPlayer::jumped-back and PtPlayer::jumped-forward are not
used within Parlatype anymore but still exist in PtPlayer.