Fix build of gnome-languages on WIN32
gnome-languages.c includes langinfo.h for nl_langinfo_l() which is not
available on WIN32. It is used to get the codeset of a locale. We don't
need the codeset, so this can be safely removed (for our purposes).

To get translations of other locales than the current one, it uses
locale_t, newlocale() and freelocale(). These are not available on WIN32
either. We don't need translations for other locales than the current
one, so this can be safely removed (again, for our purposes).

Also remove other code that we don't need and never call, especially
scanning the file system for available translations.