Help: remove ASR preferences tab description
Text and figure are outdated (version 2.1). ASR is now described in
asr-setup.page which is only installed, if Parlatype is configured
with CMU Pocketsphinx or with Mozilla DeepSpeech.

As far as I know it's not possible to add a conditional link from the
text block of the preferences page to the ASR page.