PtPlayer: fix setting initial position
Sometimes the stream's duration is not known yet when opening a file
and trying to go to the last known position.
pt_player_jump_to_position() didn't allow such a seek, although it is
actually possible, so allow seeks even if duration is -1.

When this happened, it was subsequently not possible to change the
position manually with the pointer because duration was not updated,
still -1 and a seek was prevented.