PtWaveviewer: don't emit signals on every selection change
While still dragging just draw the selection. Emit signals only on drag
end.

When clicking with shift button pressed, move that end of the selection
that is closer to the clicked pixel. If clicked inside a selection, it
will be made smaller at the end closer to the clicked pixel.